requirements.txt
requests>=2.28.0
pystray>=0.19.4
Pillow>=9.0.0
pyperclip>=1.8.2
pyinstaller>=5.0.0
configparser>=5.0.0